tidy up toggles PR
