+++ .buildkite/pipeline.yml
@@ -245,6 +245,7 @@
           - AWS_ACCESS_KEY_ID
           - AWS_SECRET_ACCESS_KEY
           - AWS_SESSION_TOKEN
+          - IS_CI=true
 
 - label: "deploy pa11y"
   branches: "main"
