+++ toggles/webapp/package.json
@@ -6,16 +6,16 @@
   "scripts": {
     "build": "ts-node ./build",
     "test": "jest",
-    "deploy": "ts-node ./deploy"
+    "deploy": "ts-node ./scripts/deploy"
   },
   "devDependencies": {
     "@aws-sdk/client-s3": "^3.39.0",
-    "@aws-sdk/client-sts": "^3.39.0",
     "@babel/preset-typescript": "^7.16.0",
     "@types/node": "12.19.0",
     "@types/yargs": "^17.0.5",
+    "@weco/ts-aws": "^0.1.0",
     "jest": "^27.3.1",
-    "ts-node": "^9.1.1",
+    "ts-node": "^10.2.1",
     "typescript": "^4.1.5"
   },
   "dependencies": {
