+++ toggles/webapp/s3-utils.ts
@@ -3,42 +3,9 @@ import {
   PutObjectCommand,
   S3Client,
 } from '@aws-sdk/client-s3';
-import { STSClient, AssumeRoleCommand } from '@aws-sdk/client-sts';
 import { Readable } from 'stream';
 import { TogglesResp } from '.';
-import { bucket, key, region } from './config';
-
-export async function getCredentials(arn?: string) {
-  const stsClient = new STSClient({ region });
-  const assumeRoleCommand = new AssumeRoleCommand({
-    RoleArn: arn ?? 'arn:aws:iam::130871440101:role/experience-ci',
-    RoleSessionName: 'ExperienceTogglesDeploySession',
-  });
-  const { Credentials, $metadata: assumeRoleMetadata } = await stsClient.send(
-    assumeRoleCommand
-  );
-  if (!Credentials)
-    throw Error(
-      `Could not assume role experience-admin with ${assumeRoleMetadata}`
-    );
-
-  return {
-    accessKeyId: Credentials.AccessKeyId!,
-    secretAccessKey: Credentials.SecretAccessKey!,
-    expiration: Credentials.Expiration,
-    sessionToken: Credentials.SessionToken,
-  };
-}
-
-export async function getS3Client() {
-  const credentials = await getCredentials();
-  const s3Client = new S3Client({
-    region,
-    credentials,
-  });
-
-  return s3Client;
-}
+import { bucket, key } from './config';
 
 // see: https://github.com/aws/aws-sdk-js-v3/issues/1877
 async function streamToString(stream: Readable): Promise<string> {
