+++ toggles/webapp/tsconfig.json
@@ -1,20 +1,9 @@
 {
   "compilerOptions": {
-    "target": "esnext",
-    "lib": ["esnext"],
-    "allowJs": false,
-    "skipLibCheck": true,
-    "strict": true,
-    "forceConsistentCasingInFileNames": true,
-    "noEmit": true,
-    "module": "commonjs",
-    "moduleResolution": "node",
     "esModuleInterop": true,
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "strictNullChecks": true,
-    "downlevelIteration": true
   },
-  "exclude": ["node_modules", "./test/**/*.ts"],
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"]
+  // This is for https://github.com/wellcomecollection/ts-aws
+  "ts-node": {
+    "skipIgnore": true
+  }
 }
